binder: read thread sequence number on every iteration when reaping
Objects are reaped and kfree'd when we know that all threads
that were active when the object became a zombie have left
the kernel. The loop that checks this was reading the active
thread sequence number at the beginning and caching it. However,
if in the 1st iteration there are no threads, it will cache ~0ULL
and all zombies will be reaped even if new ones arrive and need
to be protected. Check the sequence number on each iteration.

Bug: 36220321
Change-Id: I694ff29800cf4ab8d37cb3bb1d5508a525dd88da
Test: tested manually
Signed-off-by: Todd Kjos <tkjos@google.com>
Signed-off-by: Siqi Lin <siqilin@google.com>